pyproject.toml
[tool.poetry]
name = "pytest-requests-api-test"
version = "0.1.0"
description = "api tests with pytest and requests"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.4"
python-dotenv = "^1.0.0"
jsonschema = "^4.21.1"
pyyaml = "^6.0.1"
allure-pytest = "^2.13.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
